# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "Documentation de l'API hackapi"
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Serveur local"
    }
  ],
  "paths": {
...
    "/features/email-existence": {
      "get": {
        "summary": "Verify an email address",
        "tags": [
          "Feature - Email Existence"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "description": "The email address to verify",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "description": "Verification status"
                    },
                    "result": {
                      "type": "string",
                      "description": "Additional information about the email"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Email is required",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/features/email-spammer": {
      "get": {
        "summary": "Send multiple emails to a recipient",
        "tags": [
          "Feature - Email spammer"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "required": true,
            "description": "The recipient's email address",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "count",
            "required": true,
            "description": "Number of emails to send",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Emails sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Confirmation message with the number of emails sent"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input (missing or invalid email/count)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Error message explaining the issue"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Error message if something went wrong"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/features/check-password": {
      "get": {
        "summary": "Vérifie si un mot de passe est dans la liste des mots de passe les plus communs",
        "tags": [
          "Feature - Password check"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "password",
            "required": true,
            "description": "Le mot de passe à vérifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Résultat de la vérification du mot de passe",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Message indiquant si le mot de passe est commun ou peu commun"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Le mot de passe est requis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Erreur indiquant que le mot de passe est requis"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne du serveur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Message d'erreur pour l'échec de la récupération des mots de passe"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}